Return 3 vars from load_nifti instead of tuple
load_nifti returns a tuple and the rest of the code uses the 3rd value which is the nibabel nifti img object. Subsequent calls for the object methods fail. 
It is probably okay to change to ignore the nibabel object and instead use the data, affine, and other outputs from load_nifti directly, except for gpu_tracker.dump_streamlines uses img.header.get_zooms() which returns a 4-length tuple while load_nifti with return_voxsize=True would return a 3-length tuple; not sure of any implications to .dump_streamlines.